improvement: show Edit this page only when post is in English
Kudos to @cinnabarhorse for bringing this up in https://github.com/aavegotchi/aavegotchi-wiki/pull/102#issuecomment-773905945.

For localized pages (lanaguages that are not `en`), the Wiki leverages Crowdin rather than directly translating them on GitHub. This change shows the `✏️ Edit this page` button only for pages written in English (checking `router.query.lang` is equal to `en`).